--- app/src/main/java/nextstep/shoppingcart/domain/ProductRepository.kt
@@ -1,5 +1,0 @@
-package nextstep.shoppingcart.domain
-
-interface ProductRepository {
-    fun getAllProducts(): Products
-}
