+++ app/src/main/java/nextstep/shoppingcart/domain/repository/ProductRepository.kt
@@ -1,0 +1,7 @@
+package nextstep.shoppingcart.domain.repository
+
+import nextstep.shoppingcart.domain.model.Products
+
+interface ProductRepository {
+    fun getAllProducts(): Products
+}
