+++ app/src/test/java/nextstep/shoppingcart/model/PriceTest.kt
@@ -1,6 +1,6 @@
-package nextstep.shoppingcart
+package nextstep.shoppingcart.model
 
-import nextstep.shoppingcart.domain.Price
+import nextstep.shoppingcart.domain.model.Price
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
